+++ vtm-gdx/src/org/oscim/gdx/GdxMap.java
@@ -33,7 +33,7 @@ import org.oscim.layers.tile.vector.labeling.LabelLayer;
 import org.oscim.map.Layers;
 import org.oscim.map.Map;
 import org.oscim.renderer.MapRenderer;
-import org.oscim.theme.VtmThemes;
+import org.oscim.theme.internal.VtmThemes;
 import org.oscim.tiling.TileSource;
 import org.oscim.utils.Parameters;
 import org.slf4j.Logger;